moving function to proper file
