+++ .gitignore
@@ -807,7 +807,6 @@ packages/lib/SyncTargetRegistry.js
 packages/lib/Synchronizer.js
 packages/lib/TaskQueue.js
 packages/lib/WelcomeUtils.js
-packages/lib/accountTypeUtils.js
 packages/lib/array.js
 packages/lib/callbackUrlUtils.test.js
 packages/lib/callbackUrlUtils.js
