+++ packages/lib/utils/joplinCloud.ts
@@ -409,3 +409,16 @@ export function getPlans(stripeConfig: StripePublicConfig): Record<PlanName, Pla
 		},
 	};
 }
+
+enum AccountType {
+	Default = 0,
+	Basic = 1,
+	Pro = 2,
+}
+
+export function accountTypeToString(accountType: AccountType): string {
+	if (accountType === AccountType.Default) return 'Default';
+	if (accountType === AccountType.Basic) return 'Basic';
+	if (accountType === AccountType.Pro) return 'Pro';
+	throw new Error(`Invalid type: ${accountType}`);
+}
